backend/requirements.txt
flask
flask-cors
fal-client
python-dotenv
